Rename Trie.java to Trie
A Trie essentially functions similar to a linked list, but each node has three parts. First is the data, which is a string here. Second is a sibling, which line up horizontally. Lastly are children, which go vertically. The goal of this program is to have a Trie which functions as a dictionary of sorts, where the user enters a certain prefix, and a list of words with matching prefixes is returned and printed.